fixed meshes reloading
