+++ src/main/java/com/marginallyclever/ro3/render/renderpasses/DrawBackground.java
@@ -121,6 +121,7 @@ public class DrawBackground implements RenderPass {
     @Override
     public void dispose(GLAutoDrawable glAutoDrawable) {
         GL3 gl3 = glAutoDrawable.getGL().getGL3();
+        mesh.unload(gl3);
         shader.delete(gl3);
     }
