+++ src/main/java/com/marginallyclever/ro3/render/renderpasses/DrawBoundingBoxes.java
@@ -104,6 +104,7 @@ public class DrawBoundingBoxes implements RenderPass {
     @Override
     public void dispose(GLAutoDrawable glAutoDrawable) {
         GL3 gl3 = glAutoDrawable.getGL().getGL3();
+        mesh.unload(gl3);
         shader.delete(gl3);
     }
