+++ src/main/java/com/marginallyclever/ro3/render/renderpasses/DrawCameras.java
@@ -86,6 +86,7 @@ public class DrawCameras implements RenderPass {
     @Override
     public void dispose(GLAutoDrawable glAutoDrawable) {
         GL3 gl3 = glAutoDrawable.getGL().getGL3();
+        mesh.unload(gl3);
         shader.delete(gl3);
     }
