+++ src/main/java/com/marginallyclever/ro3/render/renderpasses/DrawDHParameters.java
@@ -83,6 +83,7 @@ public class DrawDHParameters implements RenderPass {
     @Override
     public void dispose(GLAutoDrawable glAutoDrawable) {
         GL3 gl3 = glAutoDrawable.getGL().getGL3();
+        mesh.unload(gl3);
         shader.delete(gl3);
     }
