+++ src/main/java/com/marginallyclever/ro3/render/renderpasses/DrawPoses.java
@@ -69,6 +69,7 @@ public class DrawPoses implements RenderPass {
     @Override
     public void dispose(GLAutoDrawable glAutoDrawable) {
         GL3 gl3 = glAutoDrawable.getGL().getGL3();
+        mesh.unload(gl3);
         shader.delete(gl3);
     }
